project/camera_conection/requirements.txt
python-dotenv
requests
opencv-python